Travis fiddling
Think its finally working - hurrah!
The reason for all of this travis fiddling is that the travis build wasn't seeing the scipy
installed with apt-get as python-scipy.
This is because the travis virtualenv isn't seeing the system packages.
I tried to fix this by pip installing scipy, numpy, etc., but this was fraught because
scipy takes *so long* so install from source.
Eventually found a handy fix online where we just tell the travis virtual env to use the
system site packages.