Support multiple grid tiles per convolution kernel
The gridding of a large convolution kernel is now broken up into
separate tiles, each handled in a different workgroup. This allows the
kernel function to be almost arbitrarily large, just requiring more time
(rather than more registers) to grid it. There is a performance hit
because the visibilities are now loaded multiple times.