Initial work on W stacking
The image-plane correction has been implemented (built in to
TaperDivide), and the kernel calculation has been modified to compute a
set of kernels symmetric around w=0. There is also support for computing
some parameters when there are multiple slices, but in most places
asking for more than one slice will flat-out not work. In particular
- only the first slice will be gridded and imaged
- the calculation for the size of the kernels needs re-doing