Code reorganisation with Imaging class
There is now one uber-OperationSequence (Imaging) that ties together all
the others. It does not contain any sequencing logic, but merely
connects all the buffers to the operations, and provides a thin veneer
over the interfaces of the individual operations.

It also hooks in the Scale operation and provides GPU-side zero-fills.
This improves performance as buffers are no longer being migrated back
to the CPU just to be zero-filled.

The --profile option is removed, since it never worked properly (it
included unified memory migration time), and it would be better to add
proper support for it in katsdpsigproc anyway.