Allow <4 input polarizations when not doing rotation
This is handled by splitting the code paths based on a template
parameter, which is either mueller_generator_simple or
mueller_generator_parallactic. This is a function object that computes
the Mueller matrix to apply for each timestep.

When no parallactic angle correction is in use, the feed angles are
passed as None, and only a composite matrix is passed (instead of two
matrices to go via circular). This should also make this code path more
efficient.