Handle multiple channels in the imager
Apart from convenience, this allows for more efficient access patterns
to measurement sets (because the channel is the fastest-varying
dimension). Eventually there is also the possibility of pushing some of
this into the GPU processing e.g. allowing a single invocation of
imager.py to use multiple GPUs in parallel.

There is still more work to be done to bound the memory (and temporary
disk) usage by doing the preprocessing in batches of channels rather
than all channels.

There is also a need to incorporate this into the imager report.