Apply border in which no CLEAN components can be added
Near the border of the image there are can be numeric instabilities due to
antialiasing, and placing components there can cause CLEAN not to
converge. A new --border option is added to control the size of a border
region, and the tiles used for peak-finding are restricted to (and
aligned with) the non-border areas.

The border region is not cropped from the final image. As far as I can
tell, this choice is consistent with wsclean. It allows large residuals
(such as uncleaned sources) to be spotted.

Closes SR-289.