Apply calibration solutions to visibilities
Where available, the solutions stored in telstate by katsdpcal are
loaded and applied to the visibilities before imaging. A command-line
option (-i apply=) gives control over which solutions to apply.

From a first test the results are not improved and are possibly worse,
but it's unclear whether the solutions are actually an improvement.