Do the same interpolation of cal solutions as katsdpcal
This means taking the last known value of K and B rather than
interpolating them. The idea is to match the values that were
pre-applied when solving for G. While this will nominally give less
accurate results, those inaccuracies would be partly absorbed in the
following solutions, and using a more accurate version would lead to
corrections being applied twice.

In practice, this made no qualititative difference to the one image I
tried.