Wire up PSF cutoff logic with --psf-cutoff
The PSF is now stored at full resolution for the entire imaging run, but
only the core is subtracted. This is necessary mostly because of the
limitations of katsdpsigproc.accel.Operation, which doesn't allow slots
to be resized on the fly. However, it's also useful that the imager
allocates all its GPU memory up front and so shouldn't suddenly run out
during a long run.