Rewrite get_tracks() to use katdal
The get_tracks() function examines the receptor activity sensors in
telstate and segments the data buffer into slices where receptors are
tracking. The original algorithm is quite fragile if some receptors
misbehave and end up with more segments than others due to overshoots.
This results in the infamous "inconsistent tracks" warning that causes
phase-ups and delay calibrations to fail.

Rewrite the function to use the same machinery as katdal to parse the
activity sensors. Find segments where all receptors are definitely
tracking to improve data quality.

The new algorithm is a bit more conservative, assigning fewer dumps
to tracks (especially at the start of each segment) and also merging
repeat activities. It is also slightly slower (13 ms vs 5 ms on 41 dumps
with 8 receptors). This is tolerated for the sake of robustness and
consistency with katdal.