Avoid pre-faulting the accumulation buffers
It used to take several minutes to allocate the buffers in large cases.
The buffers are now obtained directly from mmap, and are faulted in on
first being touched.

This won't work on Windows, because it doesn't use named shared buffers
or handle pickling, but we don't care about Windows.