Add sensors to the report writer
This is mostly so that it becomes easier to add more sensors later - I'm
not sure if the sensors I've put in are that critical.

This unfortunately required turning the report writer back into a class,
with the corresponding indentation change and addition of `self.` in a
bunch of places. There is now a base Task class that handles the details
of setting up sensor forwarding and dynamically choosing the Process
implementation.