Unit tests for the control module
This runs an observation through the katcp server interface. It still
needs more work to test that
- it finds sensible solutions
- it handles the different conditions that cause switching between
  buffers

and generally will need further tests to improve coverage.

There are also some structural changes to make it easier to test the
server - mostly moving code from run_cal.py into control.py so that it
can be placed under the test. The forwarding of events from master_queue
to a second asyncio queue is also eliminated in favour of using
run_in_executable to directly pull items off the master queue.