Optimise and simplify _apply
The in-place and out-of-place versions are simplified by using the out=
parameter of the np.multiply ufunc. The solution is inverted before
combining the left and right sides (less work, may be fractionally less
accurate but only at the ulp level), and the loop (which had a nasty
access pattern) is replaced by fancy indexing.