Fix and optimise wavg with numba
This includes a fix where NaNs in the data would not contribute to the
weighted sum, but the corresponding weight would still be included in
the normalisation factor, leading the returned value to be scaled down.

It proved difficult to write a completely generic numba implementation,
and even the rather complicated one I managed had noticably worse
performance than the version used here, so I've gone with implementing a
fast path (4 dimensions, averaging over time or frequency) and a
fallback. The fallback is also more efficient than the original, but
it's the numba-optimised fast path that really shines.

wavg_full_t can still benefit from similar treatment.