Further optimise wavg
The da.where implementation goes via np.choose, which is much slower
than np.where. Drop in our own implementation that uses np.where.

Also removed the vestiges of the numba-accelerated version. While it
would be possible to bring back, it's inefficient because there is no
easy way to get two numbers back from a numba ufunc (the numba support
for structured values seems weak), so it would be necessary to compute
some of the values twice. Experiments suggest that it still gives a bit
of performance improvement, but not enough to justify the effort.