Use dask diagnostics profilers as context managers
There are a number of issues with the register/unregister approach I was
using previously; it's not equivalent (see e.g. dask/dask#2551). This
meant a bit of code reorganisation to split out the body into a `_run`
method, so that it can be called from both branches of the if statement.