Remove unnecessary code from katsdpcal/__init__.py
It was a workaround for https://github.com/casacore/casacore/issues/243.
Since spead2 is no longer based on Boost.Python, the issue is no longer
of concern. It's also been fixed in casacore.

I also removed the override of the log level, since that should be up to
the user of the library, not the library itself.

I removed the version pin on python-casacore because docker-base now
provides a default version.