Remove starting time contortions
Previous solution seemed to take the second timestamp in the observation
and remove the hours minutes and seconds component from it, to produce a
starting time, `time0`.

Just take the first observation timestamp as `time0` since, for AIP's
purposes, it is merely subtracted from other timestamps
and then converted to days.