Refactor the accumulator
The old version had an issue that when a break condition was reached,
the first dump of the new target/whatever got thrown in with the
previous batch.

In the new version, the _accumulate method handles a whole observation,
rather than returning at the end of a batch. This allows it to flush to
the pipeline exactly when it needs to. It's also had a lot of pieces of
code split out into smaller functions to make it more manageable.

There will also be a slightly higher latency to flush when the
max_length is reached (it will only happen after reading the next heap),
but given that max_length is expected to be at least 100 in typical use
and that it should also not be hit in normal use, it doesn't seem worth
worrying about.