Support for multiple ingests
This is rather limited support: it has no support for the L0 product
having a different channelisation (either in bandwidth or channel width)
than the CBF stream, and large bits of the code still pull from cbf_*
telstate attributes (or indeed *writes* to such attributes). It just
fixes up the accumulator.

The tests also need to be extended to actually test the support, which
will be done in a later commit.