Improve handling of flagged data in k_fit.
In the current implementation of delaycal NaNs in the visibility data are zeroed in order to guarantee a successful FFT in the determination of coarse delays.
This strategy causes stefcal to fail if all baselines to the reference antenna are flagged/zeroed on a single channel. This failure occurs as the sum of the weighted visibilities across all baselines will be zero for this channel and thus the term  g *= np.conj(ref) / np.abs(ref) in stefcal is 0+0j/0 which raises a complex divide by zero error.

The proposed improvement interpolates across NaNs in the visibilities for the coarse delay determination, instead of simply zeroing them.

There are also a few improvements to the handling of flags in the bandpass fit in delaycal.

1.) Flags are inputted into the stefcal as zero weights. This improves the gain estimation on channels with only a few flagged baselines.
2.) Completely flagged antennas are removed from the gain estimation in stefcal (on a per channel basis). This prevents these antennas from producing NaN's in the  g[p] = gn / gd step of stefcal which then cause all subsequent estimates of g to all antennas to become NaN's.
3.) The gains of completely flagged antennas are set to NaN (on a per channel basis), to indicate that these gains are flagged.

This new strategy produces a bandpass with NaN's on completely flagged antennas. However np.unwrap and np.linalg.lstsq do not produce sensible outputs if there are NaNs in their input. Therefore a ~np.isnan mask is used to remove NaNs in the final least squares fit step of delaycal.