Proxy katdal data attribute access and chunk scans
Previously the `uv_scans` method would retrieve all data in a scan. For
full MeerKAT resolution this is impractical as scans might quickly
exceed the memory limit of a machine. This commit modifies the
KatdalAdapter to provide proxied, indexed access to the data attributes
of a Katdal object.

Providing visibility indexing capabilities allows us to select chunks of
the scan defined by a starting and ending time index. Thus, we can still
retrieve and transform parts of the scan while writing to AIPS disk.

To facilitate this, the KatdalAdapter now supports

- a `scans` method, proxying Katdal `scans`.
- a `uv_vis` method, proxying Katdal `vis`, `weight` and `flag` data
  attributes.
- a `uv_timestamps` method, proxying Katdal `timestamps` data attribute.
- a `uv_u` method, proxying Katdal `u` data attribute.
- a `uv_v` method, proxying Katdal `v` data attribute.
- a `uv_w` method, proxying Katdal `w` data attribute.

The `KatdalAdapter.uv_scans` method has been moved to the
`time_chunked_scans` function.