Improve handling of end-of-stream
1. Wait up to 5 seconds for the accumulator to stop "naturally" (due to
stop heaps in the stream).
2. Wait for all the ingests to send stop heaps before stopping the
accumulator, instead of just the first one. This will avoid data loss if
the ingests are not quite in sync.

A lot of the work was in redoing the mechanism for mocking the stream in
the test. The "pseudo-heaps" (dictionaries to be dumped straight into
the ItemGroup) have been replaced by real spead2 heaps. This new design
is based on the mocking in katsdpingest and makes it easier to send the
stop heaps etc.