Refactor plotting functions to avoid code repetition
This commit creates separate functions to
* add a frequency axis to plots of the spectrum
* determine an appropriate amplitude range for range limited plots
which can then be reused in all plots which require these functions.

Additionally it combines the previously separate functions for plotting amplitude and phase, or only amplitude into one function.

It also alters the x-axis of plots versus channel number to use the original channel numbers as opposed to the arbitrary reduced number of channels after averaging.