Improve plots in the RFI summary in the cal report
*Improve accuracy of reported % time flagged
Previously the report plotted the scan averaged flags in its plot of % time flagged as a function of baseline and channel. These plots have limited accuracy and the scan averaged flags can only return a True or False value per channel and baseline. The accuracy of this plot has been improved by calculating the sum of the flags across the time axis per scan and passing this to the report writer. This allows a more accurate determination of the %time flagged.

*Improve labels of baseline axis
The baseline axis has been sorted according to antenna separation.
The baseline axis has been labelled with antenna separation

*Improve labels of scan axis
The scan axis has been labeled with the target source in each scan.