Deal gracefully with out-of-order heaps
This handles the case where one ingest is running late relative to
another, and so dump indices are not monotonic. It makes a best effect
to find the slot where a heap is supposed to go. There are still
theoretically cases where the slot has already been flushed, or where
the delayed dump index has never been seen before, where the data will
still be lost, but this should cover the "common" case (still uncommon,
but has actually been observed in simulation).

Fixes SR-1029.