Fix bug which double applies calibration to targets in the pipeline.
Previously after processing a track the pipeline would apply all calculated solutions regardless of the tags associated with source in the processed track. However objects with a 'target' tag already have the solutions applied using `apply_inplace`.
This commit checks the tags associated with the track target and only applies solutions at the end of the track processing if the source does not have a `target` tag.