Plumb in shared_solve
It is now used to compute G, K and KCROSS solutions.

There is still a serious race condition leading to non-deterministic
behaviour: get_solns_to_apply grabs whatever solutions it finds in
telstate, which may have come from the "future" as another server runs
ahead of this one.

This passes unit tests, but has not been tested in the multi-server
scenario.