Internally store computed solutions
This will reduce the number of trips to the telescope state, but more
importantly, it fixes a race condition in split cal where requests to
telescope state might or might not pick up future solutions computed by
other servers. Now, the pipeline only retrieves solutions that it
computed itself or previously obtained through shared_solve.

Only G solutions are stored indefinitely, which shouldn't be an issue
before they're tiny. For the other solution types, only the latest value
is retained.