Refactor KCROSS-FIT
This commit unifies `kcross_fit` and `kfit` into a single function which will fit delays to cross-correlation data or auto-correlation data. It determines which type of data it is fitting based on the supplied `corrprod_lookup` parameter.

The pre-fit data averaging which was previously implemented in `kcross_fit` has been moved to `kcross_sol` in ':class: Scan'. Data weights have now been included in this averaging step.

When solveing for 'KCROSS_DIODE', the process of NaNing solutions for antennas with non-firing noise-diodes has been incorporated into the 'kcross_sol' function.