Add rfi_freq_chunks parameter
The main benefit of this change is that the parameter is properly scaled
according to the number of cal servers, whereas previously the width of
the chunks varied with the number of servers.

The default supplied in the config files is now 8 rather than 10, which
has the advantage of being a power of 2 and hence likely to be evenly
divisible by the number of cal servers. The effect hasn't been tested
but it is expected to be small.

The code in pipelineprocs for converting between origin parameter, local
version and version that goes in telstate has been generalised to allow
the table to specify arbitrary conversions.