Fix missing timestamp and dump index for lost heaps
If zero heaps arrive for a particular dump index on cal, it will
allocate a slot for it (with data_lost flags), but fails to write in the
timestamp and dump index, which may easily confuse code that works with
the timestamps.

This hopefully fixes SR-1096, but it will need further testing to
confirm.