Summarize completely flagged solutions in the report
It is difficult to pick out by eye from the plots which antennas have been completely flagged as this data is NaN'ed and hence not plotted. To improve the detectability of flagged antennas a summary of flagged antennas is included for some cal_products. These antennas are also indicated in the solution plot legend.