Produce report for observation with no tracks
If there are no tracks in an observation then the corrected data is an empty dictionary. Previously the report writer crashed if the required keys were missing from the corrected data dictionary.
The report writer now checks whether the corrected data dictionary is empty before attempting to write any relevant section of the report.