Allow number of scans in report writer buffer to be limited
This will hopefully put an end to cal-out-of-memory issues. Each time a
batch of scan summaries is received by the report writer and appended to
its lists, they are truncated to the value of --max-scans (if given). By
default it is unbounded. Also added sensors to track the number of scan
summaries received (ever) and the current buffer size.

This is still quite as bounded as it could be, because
(a) it does not bound the number of summaries in the multiprocessing
queue, which theoretically could grow if the report writer falls behind;
and
(b) it does not bound the number of summaries sent in each batch through
the queue (other than that it cannot be more than the number of
timestamps in the buffer).

However, it should still be good enough to handle the normal failure
case, where memory usage grows without bound over many hours.