Restrict PSF patch size
In short observations with noisy PSFs, the heuristic to choose the patch
size can end up taking the entire image as the "patch". The edges are
pretty suspect and ideally shouldn't be used in CLEAN subtraction. With
this commit, only the central 50% on each axis (by default - tunable) is
a candidate for the CLEAN patch.

Closes SR-1263.