More accurate transfer of weights from simulator
The simulator attempted to reverse the synthesis of weights done by
katdal, but this is getting more complicated in the future and somewhat
impractical to reverse. Instead, send every weight as a float32. This
will increase the bandwidth beyond what the real system uses, but
ensures the weights produced by katdal are faithfully transferred into
the calibration (and should match those that went in the first time
around).