Implement continuum_factor support
This wires calprocs.wavg_flags_f into control, so that it is used to do
reduction of flags in streams with continuum_factor > 1.

The wavg_flags_f function was also made more generic so that it can be
used with different numbers of axes without requiring lots of axis
mangling by the caller (it's done by the function instead).

Relates to SR-1351.