Write report with no reference antenna
The procedure to select a reference antenna is only triggered in the pipeline when there is a calibrator tag. For observations with only 'target' tags or no tags a reference antenna is not selected. Previously this caused the report writer to crash as refant_index would be set to 'None'. This pull request allows a report to be written even if no reference antenna has been selected.