Get correct Obit revision when logging.
Obit is version controlled with subversion and logs the subvsersion revision number
of each task at runtime. This information can get lost if Obit is installed from
a working directory that is not svn controlled. This will retrieve the Obit source
using subversion so that the revision number is not lost in the Obit logs.